Fix up some permission loopholes.
